requirements.txt
# All the libraries I installed
# "pip install -r requirements.txt" -> to install all the libraries

requests
openai
dotenv
schedule
snowflake-connector-python